Reuse jobowners
